pyproject.toml
[project]
name = "ai-tools"
version = "0.1.0"
author = "Izzaz Iskandar"
email = "izzaz@time.com.my"
description = "AI Tools for Izzaz"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "ipykernel>=6.29.5",
    "openai>=1.75.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
]
[project.scripts]
hello = "example.hello"
